try to fix a problem: when a user left from group bot needs to analize and send a message this event for all admins. But this handler need rewrite, bcs idk why is is not working. we need to find more solutions this problem

+++ group_events/events.py
@@ -45,7 +45,7 @@ async def joined_user(event: ChatMemberUpdated):
 
 @events.chat_member()
 async def left_user(event: ChatMemberUpdated):
-    # if event.status == "left":
+    if event.old_chat_member.status == 'left':
         await bot.delete_message(event.chat.id, 
                                 event.message.message_id)
         for admin in await bot.get_chat_administrators(event.chat.id):
